requirements.txt
Django==4.1
flake8==5.0.4 
mccabe==0.7.0 
pycodestyle==2.9.1 
pyflakes==2.5.0
black==22.6.0 
click==8.1.3 
colorama==0.4.5 
mypy-extensions==0.4.3 
pathspec==0.9.0 
platformdirs==2.5.2 
tomli==2.0.1
django-braces==1.15.0 
django-model-utils==4.2.0 
django-star-ratings==0.9.2 
swapper==1.3.0
Pillow==9.2.0
